Moved the squares array to history (an array of arrays) in order to watch every step on the game
